Convert lk_ball to tensor-param dispatch.
Convert lk_ball to tensor-param dispatch. Factor lk_ball torch_ops from
compiled layer into torch_ops submodule.
