+++ tmol/score/lk_ball/potentials/compiled.cc
@@ -3,6 +3,7 @@
 #include <torch/torch.h>
 
 #include <tmol/score/common/dispatch.cpu.impl.hh>
+#include <tmol/utility/function_dispatch/pybind.hh>
 
 #include <tmol/score/ljlk/potentials/params.pybind.hh>
 #include "../../bonded_atom.pybind.hh"
@@ -16,30 +17,34 @@ namespace score {
 namespace lk_ball {
 namespace potentials {
 
-template <typename Real, typename Int>
+template <tmol::Device Dev, typename Real, typename Int>
 void bind_dispatch(pybind11::module& m) {
   using namespace pybind11::literals;
+  using namespace tmol::utility::function_dispatch;
 
-  m.def(
+  add_dispatch_impl<Dev, Real>(
+      m,
       "attached_waters_forward",
-      attached_waters<Real, Int, tmol::Device::CPU, 4>::forward,
+      attached_waters<Real, Int, Dev, 4>::forward,
       "coords"_a,
       "indexed_bonds"_a,
       "atom_types"_a,
       "global_params"_a);
 
-  m.def(
+  add_dispatch_impl<Dev, Real>(
+      m,
       "attached_waters_backward",
-      attached_waters<Real, Int, tmol::Device::CPU, 4>::backward,
+      attached_waters<Real, Int, Dev, 4>::backward,
       "dE_dW"_a,
       "coords"_a,
       "indexed_bonds"_a,
       "atom_types"_a,
       "global_params"_a);
 
-  m.def(
+  add_dispatch_impl<Dev, Real>(
+      m,
       "lk_ball_V",
-      &LKBallDispatch<common::NaiveDispatch, tmol::Device::CPU, Real, Int>::V,
+      &LKBallDispatch<common::NaiveDispatch, Dev, Real, Int>::V,
       "coords_i"_a,
       "coords_j"_a,
       "waters_i"_a,
@@ -51,9 +56,10 @@ void bind_dispatch(pybind11::module& m) {
       "global_lkb_params"_a,
       "global_lj_params"_a);
 
-  m.def(
+  add_dispatch_impl<Dev, Real>(
+      m,
       "lk_ball_dV",
-      &LKBallDispatch<common::NaiveDispatch, tmol::Device::CPU, Real, Int>::dV,
+      &LKBallDispatch<common::NaiveDispatch, Dev, Real, Int>::dV,
       "coords_i"_a,
       "coords_j"_a,
       "waters_i"_a,
@@ -69,7 +75,7 @@ void bind_dispatch(pybind11::module& m) {
 PYBIND11_MODULE(TORCH_EXTENSION_NAME, m) {
   using namespace pybind11::literals;
 
-  bind_dispatch<float, int64_t>(m);
+  bind_dispatch<tmol::Device::CPU, float, int64_t>(m);
 }
 
 }  // namespace potentials
